refactor(refs T33194, T19759): Move Files for a better structure
