+++ src/components/core/DpEditor/modals/DpLinkModal.vue
@@ -46,11 +46,11 @@
 </template>
 
 <script>
-import DpButtonRow from '../../DpButtonRow/DpButtonRow'
-import DpCheckbox from '../../DpCheckbox/DpCheckbox'
-import DpInput from '../../DpInput/DpInput'
-import DpModal from '../../DpModal/DpModal'
-import { dpValidateMixin } from '../../../mixins'
+import DpButtonRow from '../../../DpButtonRow/DpButtonRow'
+import DpCheckbox from '../../../DpCheckbox/DpCheckbox'
+import DpInput from '../../../DpInput/DpInput'
+import DpModal from '../../../DpModal/DpModal'
+import { dpValidateMixin } from '../../../../mixins'
 
 export default {
   name: 'DpLinkModal',
